pyproject.toml
[tool.poetry]
name = "rfd-notify"
version = "1.0.0"
description = "get notified of deals based on regex"
authors = ["Dave Gallant <dave.gallant@gmail.com>"]
readme = "README.md"
packages = [{include = "rfd_notify"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
pyyaml = "^6.0"
loguru = "^0.6.0"
apprise = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^23.1.0"
pylint = "^2.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
